Complete JWT-secured cart module with user-specific access and DTO-based responses

+++ Backend/Electronic-shopping-backend/src/main/java/com/backend/user/controllers/UserController.java
@@ -112,7 +112,7 @@ public class UserController {
 	
 	@GetMapping("/{id}")
 	public ResponseEntity<ApiResponse<UserDetailsResponseDTO>> getUserById(
-	        @PathVariable Integer id) {
+	        @PathVariable Long id) {
 
 	    UserDetailsResponseDTO data = userService.getUserById(id);
 
@@ -124,7 +124,7 @@ public class UserController {
 	
 	@PutMapping("/{id}")
 	public ResponseEntity<ApiResponse<UserDetailsResponseDTO>> updateUser(
-	        @PathVariable Integer id,
+	        @PathVariable Long id,
 	        @Valid @RequestBody UpdateUserDTO dto) {
 
 	    UserDetailsResponseDTO data = userService.updateUser(id, dto);
